If we fail to check a repo, report that and keep going
Previously, it would simply die with an error message (to avoid giving
incorrect data about what vulnerabilities exist, since we were not able
to scan a particular repo). Instead, this reports that as a HIGH
severity issue and lets the scan continue so that we can still access
the other repos' results, while still knowing about this problem.